backend/requirements.txt
asgiref==3.11.1
Django==5.1.15
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
psycopg==3.3.3
psycopg-binary==3.3.3
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0
